Q1: host quantizer in l3_transforms - buildQuantTables (luma rows of svt_av1_build_quantizer md_config_process.c:106-135 at sharpness=0; dc/ac qtx lookups inv_transforms.c:3467/:3484 verbatim tables :3412/:3357; invert_quant :3516; qzbin :3501), defaultScan4x4 (svt_aom_init_iscan formula coefficients.c:345-363), quantizeFp4x4 (quantize_fp_helper_c full_loop.c:222 at log_scale 0) + quantizeB4x4 (svt_aom_quantize_b_c :31); tests vs gate lines qtab_q100 + qfp_*/qb_* at q{0,1,100,200,255} (fp mutation-checked: shift+1 breaks 3/5 qindices; b-path RED: missing wt=1<<AOM_QM_BITS multiplies, all 5 fail, then green); layer decision: quantizer lives in l3 (no new layer) - AGENTS.md updated; RIDE-ALONG: sad4x4/sad8x8 kernels now take unsigned-int strides + write unsigned-int out (host signatures matched); green-before 7/7 = green-after 7/7, no behavior change; GATE 71/71

+++ src/l5_motion/motion.cpp
@@ -20,16 +20,17 @@ std::uint32_t sad8x8(const std::uint8_t* src, std::uint32_t srcStride, const std
 
 std::string sad8x8CuSource() {
     return R"CUDA(
-extern "C" __global__ void sad8x8_kernel(const unsigned char* src, const int* srcStride,
-                                         const unsigned char* ref, const int* refStride, int* out) {
-    int sad = 0;
-    for (int r = 0; r < 8; ++r) {
-        for (int c = 0; c < 8; ++c) {
+extern "C" __global__ void sad8x8_kernel(const unsigned char* src, const unsigned int* srcStride,
+                                         const unsigned char* ref, const unsigned int* refStride,
+                                         unsigned int* out) {
+    unsigned int sad = 0;
+    for (unsigned int r = 0; r < 8; ++r) {
+        for (unsigned int c = 0; c < 8; ++c) {
             int d = (int)src[c] - (int)ref[c];
             if (d < 0) {
                 d = -d;
             }
-            sad += d;
+            sad += (unsigned int)d;
         }
         src += *srcStride;
         ref += *refStride;
@@ -58,16 +59,17 @@ std::uint32_t sad4x4(const std::uint8_t* src, std::uint32_t srcStride, const std
 
 std::string sad4x4CuSource() {
     return R"CUDA(
-extern "C" __global__ void sad4x4_kernel(const unsigned char* src, const int* srcStride,
-                                         const unsigned char* ref, const int* refStride, int* out) {
-    int sad = 0;
-    for (int r = 0; r < 4; ++r) {
-        for (int c = 0; c < 4; ++c) {
+extern "C" __global__ void sad4x4_kernel(const unsigned char* src, const unsigned int* srcStride,
+                                         const unsigned char* ref, const unsigned int* refStride,
+                                         unsigned int* out) {
+    unsigned int sad = 0;
+    for (unsigned int r = 0; r < 4; ++r) {
+        for (unsigned int c = 0; c < 4; ++c) {
             int d = (int)src[c] - (int)ref[c];
             if (d < 0) {
                 d = -d;
             }
-            sad += d;
+            sad += (unsigned int)d;
         }
         src += *srcStride;
         ref += *refStride;
